[ALPS04575136] WMT: fix coverity for printf format
[Solution]
Fix printf worong type for wmt_lib, wmt_func, wmt_dbg,
stp_core, and hif_sdio.c

MTK-Commit-Id: 7cc0ebee08f862b472b71c5b0d39deab762dee87

CR-Id: ALPS04575136
Feature: [Module]WMT Driver

+++ common_main/linux/hif_sdio.c
@@ -347,7 +347,8 @@ static _osal_inline_ INT32 hif_sdio_deep_sleep_info_dmp(MTK_WCN_HIF_SDIO_DS_INFO
 
 	mutex_lock(&p_ds_info->lock);
 	HIF_SDIO_DBG_FUNC("p_ds_info: 0x%08x, chipid:0x%x, reg_offset:0x%x, value:0x%x\n",
-			   p_ds_info, p_ds_info->chip_id, p_ds_info->reg_offset, p_ds_info->value);
+			(unsigned int)p_ds_info, p_ds_info->chip_id, p_ds_info->reg_offset,
+			p_ds_info->value);
 
 	for (i = 0; i < ctl_info_array_size; i++) {
 		ctl_info = &p_ds_info->clt_info[i];
